dev-requirements.txt
coverage==3.7.1
#coverage==4.1b2
gocept.testing==1.10
nose==1.3.0
numpydoc==0.5
sphinxcontrib-exceltable==0.1
Sphinx==1.3.1
versioneer==0.13
wheel==0.24.0
tox==2.0.1
#pandas==0.13.1

